Moved sdk_new_node_caught_up fixture to top level
