+++ plenum/test/recorder/test_replay_on_new_node.py
@@ -3,8 +3,6 @@ from plenum.common.config_util import getConfigOnce
 
 from plenum.test.recorder.helper import create_replayable_node_and_check, \
     reload_modules_for_replay, get_replayable_node_class
-from plenum.test.node_catchup.conftest import sdk_node_created_after_some_txns, \
-    sdk_node_set_with_node_added_after_some_txns, sdk_new_node_caught_up
 
 TestRunningTimeLimitSec = 200
