+++ plenum/test/test_current_state_propagation.py
@@ -1,6 +1,3 @@
-from plenum.test.node_catchup.conftest import whitelist, \
-    sdk_node_set_with_node_added_after_some_txns, \
-    sdk_new_node_caught_up, sdk_node_created_after_some_txns
 from plenum.test.pool_transactions.helper import \
     disconnect_node_and_ensure_disconnected
 from plenum.test.test_node import checkNodesConnected
